refactor: move calc functions into new legend logic file
